Use retained MQTT topics to know containers
When connecting to MQTT, we will now get the retained message topics for
all containers pushed by HA-Dockermon on the base topic set.

This allows us to see which topics no longer match a container on the
system, and can then add that to the list to be marked as deleted in
time.

This is an attempt to fix #65